+++ examples/arrow-runtime/bin/start-agent.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-export PYTHONPATH=`pwd`/:${PYTHONPATH}
+export PYTHONPATH=`pwd`/:${PYTHONPATH}:`pwd`/../../
 
-cd ./arrow/agent/; nohup python ./arrow_agent.py ./conf/agent.conf </dev/null >../../log/agent.stdout.stderr &
-
+cd ./arrow/agent/; nohup python ./arrow_agent.py ../../conf/agent.conf </dev/null >../../log/agent.stdout.stderr &